requirements.txt
torch
torchvision
transformers
numpy
pillow
matplotlib
pandas
